Clean up & refactored some names to be more descriptive
