+++ tests/Container.spec.js
@@ -6,8 +6,8 @@
  * LICENSE file in the root directory of this source tree.
  */
 import Container 			from '../src/Container';
-import MochComponent 		from './mochs/MochComponent';
-import MochComponentTwo 	from './mochs/MochComponentTwo';
+import MochComponent 		from './mochs/MockComponent';
+import MochComponentTwo 	from './mochs/MockComponentTwo';
 
 let container = null;
